+++ src/scan/mod.rs
@@ -15,16 +15,26 @@ use crate::{IntRect, Rect};
 
 use crate::blitter::Blitter;
 
-pub fn fill_rect(rect: &Rect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) -> Option<()> {
-    fill_int_rect(&rect.round()?, clip, blitter)
+pub fn fill_rect(rect: &Rect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) {
+    if let Some(rect) = rect.round() {
+        fill_int_rect(&rect, clip, blitter);
+    }
 }
 
-fn fill_int_rect(rect: &IntRect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) -> Option<()> {
-    let rect = rect.intersect(&clip.to_int_rect())?.to_screen_int_rect()?;
+fn fill_int_rect(rect: &IntRect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) {
+    let rect = match rect.intersect(&clip.to_int_rect()) {
+        Some(v) => v,
+        None => return, // everything was clipped out
+    };
+
+    let rect = match rect.to_screen_int_rect() {
+        Some(v) => v,
+        None => return,
+    };
+
     blitter.blit_rect(&rect);
-    Some(())
 }
 
-pub fn fill_rect_aa(rect: &Rect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) -> Option<()> {
-    hairline_aa::fill_rect(rect, clip, blitter)
+pub fn fill_rect_aa(rect: &Rect, clip: &ScreenIntRect, blitter: &mut dyn Blitter) {
+    hairline_aa::fill_rect(rect, clip, blitter);
 }